@@ -1,28 +1,25 @@
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
-import 'package:outlook/responsive.dart';
 import 'package:websafe_svg/websafe_svg.dart';
 
 import '../constants.dart';
 import '../extensions.dart';
+import '../responsive.dart';
 import 'side_menu_item.dart';
 import 'tags.dart';
 
-import 'package:flutter/foundation.dart' show kIsWeb;
-
 class SideMenu extends StatelessWidget {
-  const SideMenu({
-    Key key,
-  }) : super(key: key);
+  const SideMenu({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Container(
       height: double.infinity,
-      padding: EdgeInsets.only(top: kIsWeb ? kDefaultPadding : 0),
+      padding: const EdgeInsets.only(top: kIsWeb ? kDefaultPadding : 0),
       color: kBgLightColor,
       child: SafeArea(
         child: SingleChildScrollView(
-          padding: EdgeInsets.symmetric(horizontal: kDefaultPadding),
+          padding: const EdgeInsets.symmetric(horizontal: kDefaultPadding),
           child: Column(
             children: [
               Row(
@@ -31,15 +28,15 @@ class SideMenu extends StatelessWidget {
                     "assets/images/Logo Outlook.png",
                     width: 46,
                   ),
-                  Spacer(),
+                  const Spacer(),
                   // We don't want to show this close button on Desktop mood
-                  if (!Responsive.isDesktop(context)) CloseButton(),
+                  if (!Responsive.isDesktop(context)) const CloseButton(),
                 ],
               ),
-              SizedBox(height: kDefaultPadding),
+              const SizedBox(height: kDefaultPadding),
               FlatButton.icon(
                 minWidth: double.infinity,
-                padding: EdgeInsets.symmetric(
+                padding: const EdgeInsets.symmetric(
                   vertical: kDefaultPadding,
                 ),
                 shape: RoundedRectangleBorder(
@@ -47,19 +44,19 @@ class SideMenu extends StatelessWidget {
                 ),
                 color: kPrimaryColor,
                 onPressed: () {},
-                icon: WebsafeSvg.asset("assets/Icons/Edit.svg", width: 16),
-                label: Text(
+                icon: WebsafeSvg.asset("assets/icons/Edit.svg", width: 16),
+                label: const Text(
                   "New message",
                   style: TextStyle(color: Colors.white),
                 ),
               ).addNeumorphism(
                 topShadowColor: Colors.white,
-                bottomShadowColor: Color(0xFF234395).withOpacity(0.2),
+                bottomShadowColor: const Color(0xFF234395).withOpacity(0.2),
               ),
-              SizedBox(height: kDefaultPadding),
+              const SizedBox(height: kDefaultPadding),
               FlatButton.icon(
                 minWidth: double.infinity,
-                padding: EdgeInsets.symmetric(
+                padding: const EdgeInsets.symmetric(
                   vertical: kDefaultPadding,
                 ),
                 shape: RoundedRectangleBorder(
@@ -67,44 +64,44 @@ class SideMenu extends StatelessWidget {
                 ),
                 color: kBgDarkColor,
                 onPressed: () {},
-                icon: WebsafeSvg.asset("assets/Icons/Download.svg", width: 16),
-                label: Text(
+                icon: WebsafeSvg.asset("assets/icons/Download.svg", width: 16),
+                label: const Text(
                   "Get messages",
                   style: TextStyle(color: kTextColor),
                 ),
               ).addNeumorphism(),
-              SizedBox(height: kDefaultPadding * 2),
+              const SizedBox(height: kDefaultPadding * 2),
               // Menu Items
               SideMenuItem(
                 press: () {},
                 title: "Inbox",
-                iconSrc: "assets/Icons/Inbox.svg",
+                iconSrc: "assets/icons/Inbox.svg",
                 isActive: true,
                 itemCount: 3,
               ),
               SideMenuItem(
                 press: () {},
                 title: "Sent",
-                iconSrc: "assets/Icons/Send.svg",
+                iconSrc: "assets/icons/Send.svg",
                 isActive: false,
               ),
               SideMenuItem(
                 press: () {},
                 title: "Drafts",
-                iconSrc: "assets/Icons/File.svg",
+                iconSrc: "assets/icons/File.svg",
                 isActive: false,
               ),
               SideMenuItem(
                 press: () {},
                 title: "Deleted",
-                iconSrc: "assets/Icons/Trash.svg",
+                iconSrc: "assets/icons/Trash.svg",
                 isActive: false,
                 showBorder: false,
               ),
 
-              SizedBox(height: kDefaultPadding * 2),
+              const SizedBox(height: kDefaultPadding * 2),
               // Tags
-              Tags(),
+              const Tags(),
             ],
           ),
         ),
