@@ -4,9 +4,7 @@ import 'package:websafe_svg/websafe_svg.dart';
 import '../constants.dart';
 
 class Tags extends StatelessWidget {
-  const Tags({
-    Key key,
-  }) : super(key: key);
+  const Tags({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -14,23 +12,23 @@ class Tags extends StatelessWidget {
       children: [
         Row(
           children: [
-            WebsafeSvg.asset("assets/Icons/Angle down.svg", width: 16),
-            SizedBox(width: kDefaultPadding / 4),
-            WebsafeSvg.asset("assets/Icons/Markup.svg", width: 20),
-            SizedBox(width: kDefaultPadding / 2),
+            WebsafeSvg.asset("assets/icons/Angle down.svg", width: 16),
+            const SizedBox(width: kDefaultPadding / 4),
+            WebsafeSvg.asset("assets/icons/Markup.svg", width: 20),
+            const SizedBox(width: kDefaultPadding / 2),
             Text(
               "Tags",
               style: Theme.of(context)
                   .textTheme
                   .button
-                  .copyWith(color: kGrayColor),
+                  ?.copyWith(color: kGrayColor),
             ),
-            Spacer(),
+            const Spacer(),
             MaterialButton(
-              padding: EdgeInsets.all(10),
+              padding: const EdgeInsets.all(10),
               minWidth: 40,
               onPressed: () {},
-              child: Icon(
+              child: const Icon(
                 Icons.add,
                 color: kGrayColor,
                 size: 20,
@@ -38,17 +36,17 @@ class Tags extends StatelessWidget {
             )
           ],
         ),
-        SizedBox(height: kDefaultPadding / 2),
-        buildTag(context, color: Color(0xFF23CF91), title: "Design"),
-        buildTag(context, color: Color(0xFF3A6FF7), title: "Work"),
-        buildTag(context, color: Color(0xFFF3CF50), title: "Friends"),
-        buildTag(context, color: Color(0xFF8338E1), title: "Family"),
+        const SizedBox(height: kDefaultPadding / 2),
+        buildTag(context, color: const Color(0xFF23CF91), title: "Design"),
+        buildTag(context, color: const Color(0xFF3A6FF7), title: "Work"),
+        buildTag(context, color: const Color(0xFFF3CF50), title: "Friends"),
+        buildTag(context, color: const Color(0xFF8338E1), title: "Family"),
       ],
     );
   }
 
   InkWell buildTag(BuildContext context,
-      {@required Color color, @required String title}) {
+      {required Color color, required String title}) {
     return InkWell(
       onTap: () {},
       child: Padding(
@@ -56,17 +54,17 @@ class Tags extends StatelessWidget {
         child: Row(
           children: [
             WebsafeSvg.asset(
-              "assets/Icons/Markup filled.svg",
+              "assets/icons/Markup filled.svg",
               height: 18,
               color: color,
             ),
-            SizedBox(width: kDefaultPadding / 2),
+            const SizedBox(width: kDefaultPadding / 2),
             Text(
               title,
               style: Theme.of(context)
                   .textTheme
                   .button
-                  .copyWith(color: kGrayColor),
+                  ?.copyWith(color: kGrayColor),
             ),
           ],
         ),
