@@ -1,13 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:outlook/responsive.dart';
 import 'package:websafe_svg/websafe_svg.dart';
 
 import '../../../constants.dart';
+import '../../../responsive.dart';
 
 class Header extends StatelessWidget {
-  const Header({
-    Key key,
-  }) : super(key: key);
+  const Header({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -16,55 +14,55 @@ class Header extends StatelessWidget {
       child: Row(
         children: [
           // We need this back button on mobile only
-          if (Responsive.isMobile(context)) BackButton(),
+          if (Responsive.isMobile(context)) const BackButton(),
           IconButton(
             icon: WebsafeSvg.asset(
-              "assets/Icons/Trash.svg",
+              "assets/icons/Trash.svg",
               width: 24,
             ),
             onPressed: () {},
           ),
           IconButton(
             icon: WebsafeSvg.asset(
-              "assets/Icons/Reply.svg",
+              "assets/icons/Reply.svg",
               width: 24,
             ),
             onPressed: () {},
           ),
           IconButton(
             icon: WebsafeSvg.asset(
-              "assets/Icons/Reply all.svg",
+              "assets/icons/Reply all.svg",
               width: 24,
             ),
             onPressed: () {},
           ),
           IconButton(
             icon: WebsafeSvg.asset(
-              "assets/Icons/Transfer.svg",
+              "assets/icons/Transfer.svg",
               width: 24,
             ),
             onPressed: () {},
           ),
-          Spacer(),
+          const Spacer(),
           // We don't need print option on mobile
           if (Responsive.isDesktop(context))
             IconButton(
               icon: WebsafeSvg.asset(
-                "assets/Icons/Printer.svg",
+                "assets/icons/Printer.svg",
                 width: 24,
               ),
               onPressed: () {},
             ),
           IconButton(
             icon: WebsafeSvg.asset(
-              "assets/Icons/Markup.svg",
+              "assets/icons/Markup.svg",
               width: 24,
             ),
             onPressed: () {},
           ),
           IconButton(
             icon: WebsafeSvg.asset(
-              "assets/Icons/More vertical.svg",
+              "assets/icons/More vertical.svg",
               width: 24,
             ),
             onPressed: () {},
