@@ -1,23 +1,19 @@
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
-import 'package:outlook/components/side_menu.dart';
-import 'package:outlook/models/Email.dart';
-import 'package:outlook/responsive.dart';
-import 'package:outlook/screens/email/email_screen.dart';
 import 'package:websafe_svg/websafe_svg.dart';
 
+import '../../../components/side_menu.dart';
 import '../../../constants.dart';
+import '../../../models/email.dart';
+import '../../../responsive.dart';
+import '../../email/email_screen.dart';
 import 'email_card.dart';
 
-import 'package:flutter/foundation.dart' show kIsWeb;
-
 class ListOfEmails extends StatefulWidget {
-  // Press "Command + ."
-  const ListOfEmails({
-    Key key,
-  }) : super(key: key);
+  const ListOfEmails({Key? key}) : super(key: key);
 
   @override
-  _ListOfEmailsState createState() => _ListOfEmailsState();
+  State<ListOfEmails> createState() => _ListOfEmailsState();
 }
 
 class _ListOfEmailsState extends State<ListOfEmails> {
@@ -27,11 +23,11 @@ class _ListOfEmailsState extends State<ListOfEmails> {
     return Scaffold(
       key: _scaffoldKey,
       drawer: ConstrainedBox(
-        constraints: BoxConstraints(maxWidth: 250),
-        child: SideMenu(),
+        constraints: const BoxConstraints(maxWidth: 250),
+        child: const SideMenu(),
       ),
       body: Container(
-        padding: EdgeInsets.only(top: kIsWeb ? kDefaultPadding : 0),
+        padding: const EdgeInsets.only(top: kIsWeb ? kDefaultPadding : 0),
         color: kBgDarkColor,
         child: SafeArea(
           right: false,
@@ -47,12 +43,13 @@ class _ListOfEmailsState extends State<ListOfEmails> {
                     // Also we want to hide this menu icon on desktop
                     if (!Responsive.isDesktop(context))
                       IconButton(
-                        icon: Icon(Icons.menu),
+                        icon: const Icon(Icons.menu),
                         onPressed: () {
-                          _scaffoldKey.currentState.openDrawer();
+                          _scaffoldKey.currentState?.openDrawer();
                         },
                       ),
-                    if (!Responsive.isDesktop(context)) SizedBox(width: 5),
+                    if (!Responsive.isDesktop(context))
+                      const SizedBox(width: 5),
                     Expanded(
                       child: TextField(
                         onChanged: (value) {},
@@ -64,11 +61,11 @@ class _ListOfEmailsState extends State<ListOfEmails> {
                             padding: const EdgeInsets.all(
                                 kDefaultPadding * 0.75), //15
                             child: WebsafeSvg.asset(
-                              "assets/Icons/Search.svg",
+                              "assets/icons/Search.svg",
                               width: 24,
                             ),
                           ),
-                          border: OutlineInputBorder(
+                          border: const OutlineInputBorder(
                             borderRadius: BorderRadius.all(Radius.circular(10)),
                             borderSide: BorderSide.none,
                           ),
@@ -78,35 +75,35 @@ class _ListOfEmailsState extends State<ListOfEmails> {
                   ],
                 ),
               ),
-              SizedBox(height: kDefaultPadding),
+              const SizedBox(height: kDefaultPadding),
               Padding(
                 padding:
                     const EdgeInsets.symmetric(horizontal: kDefaultPadding),
                 child: Row(
                   children: [
                     WebsafeSvg.asset(
-                      "assets/Icons/Angle down.svg",
+                      "assets/icons/Angle down.svg",
                       width: 16,
                       color: Colors.black,
                     ),
-                    SizedBox(width: 5),
-                    Text(
+                    const SizedBox(width: 5),
+                    const Text(
                       "Sort by date",
                       style: TextStyle(fontWeight: FontWeight.w500),
                     ),
-                    Spacer(),
+                    const Spacer(),
                     MaterialButton(
                       minWidth: 20,
                       onPressed: () {},
                       child: WebsafeSvg.asset(
-                        "assets/Icons/Sort.svg",
+                        "assets/icons/Sort.svg",
                         width: 16,
                       ),
                     ),
                   ],
                 ),
               ),
-              SizedBox(height: kDefaultPadding),
+              const SizedBox(height: kDefaultPadding),
               Expanded(
                 child: ListView.builder(
                   itemCount: emails.length,
@@ -118,8 +115,7 @@ class _ListOfEmailsState extends State<ListOfEmails> {
                       Navigator.push(
                         context,
                         MaterialPageRoute(
-                          builder: (context) =>
-                              EmailScreen(email: emails[index]),
+                          builder: (context) => const EmailScreen(),
                         ),
                       );
                     },
