@@ -5,21 +5,21 @@ import '../constants.dart';
 import 'counter_badge.dart';
 
 class SideMenuItem extends StatelessWidget {
-  const SideMenuItem({
-    Key key,
-    this.isActive,
-    this.isHover = false,
-    this.itemCount,
-    this.showBorder = true,
-    @required this.iconSrc,
-    @required this.title,
-    @required this.press,
-  }) : super(key: key);
-
   final bool isActive, isHover, showBorder;
   final int itemCount;
   final String iconSrc, title;
   final VoidCallback press;
+
+  const SideMenuItem({
+    Key? key,
+    this.isHover = false,
+    this.itemCount = 0,
+    this.showBorder = true,
+    required this.isActive,
+    required this.iconSrc,
+    required this.title,
+    required this.press,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -31,16 +31,16 @@ class SideMenuItem extends StatelessWidget {
           children: [
             (isActive || isHover)
                 ? WebsafeSvg.asset(
-                    "assets/Icons/Angle right.svg",
+                    "assets/icons/Angle right.svg",
                     width: 15,
                   )
-                : SizedBox(width: 15),
-            SizedBox(width: kDefaultPadding / 4),
+                : const SizedBox(width: 15),
+            const SizedBox(width: kDefaultPadding / 4),
             Expanded(
               child: Container(
-                padding: EdgeInsets.only(bottom: 15, right: 5),
+                padding: const EdgeInsets.only(bottom: 15, right: 5),
                 decoration: showBorder
-                    ? BoxDecoration(
+                    ? const BoxDecoration(
                         border: Border(
                           bottom: BorderSide(color: Color(0xFFDFE2EF)),
                         ),
@@ -53,16 +53,16 @@ class SideMenuItem extends StatelessWidget {
                       height: 20,
                       color: (isActive || isHover) ? kPrimaryColor : kGrayColor,
                     ),
-                    SizedBox(width: kDefaultPadding * 0.75),
+                    const SizedBox(width: kDefaultPadding * 0.75),
                     Text(
                       title,
-                      style: Theme.of(context).textTheme.button.copyWith(
+                      style: Theme.of(context).textTheme.button?.copyWith(
                             color:
                                 (isActive || isHover) ? kTextColor : kGrayColor,
                           ),
                     ),
-                    Spacer(),
-                    if (itemCount != null) CounterBadge(count: itemCount)
+                    const Spacer(),
+                    if (itemCount != 0) CounterBadge(count: itemCount)
                   ],
                 ),
               ),
