@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:outlook/constants.dart';
-import 'package:outlook/screens/main/main_screen.dart';
+
+import 'screens/main/main_screen.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -14,7 +16,7 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       title: 'Flutter Demo',
       theme: ThemeData(),
-      home: MainScreen(),
+      home: const MainScreen(),
     );
   }
 }
