@@ -1,34 +1,34 @@
 import 'package:flutter/material.dart';
-import 'package:outlook/models/Email.dart';
 import 'package:websafe_svg/websafe_svg.dart';
 
 import '../../../constants.dart';
 import '../../../extensions.dart';
+import '../../../models/email.dart';
 
 class EmailCard extends StatelessWidget {
-  const EmailCard({
-    Key key,
-    this.isActive = true,
-    this.email,
-    this.press,
-  }) : super(key: key);
-
   final bool isActive;
   final Email email;
   final VoidCallback press;
+
+  const EmailCard({
+    Key? key,
+    this.isActive = true,
+    required this.email,
+    required this.press,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     //  Here the shadow is not showing properly
     return Padding(
-      padding: EdgeInsets.symmetric(
+      padding: const EdgeInsets.symmetric(
           horizontal: kDefaultPadding, vertical: kDefaultPadding / 2),
       child: InkWell(
         onTap: press,
         child: Stack(
           children: [
             Container(
-              padding: EdgeInsets.all(kDefaultPadding),
+              padding: const EdgeInsets.all(kDefaultPadding),
               decoration: BoxDecoration(
                 color: isActive ? kPrimaryColor : kBgDarkColor,
                 borderRadius: BorderRadius.circular(15),
@@ -44,7 +44,7 @@ class EmailCard extends StatelessWidget {
                           backgroundImage: AssetImage(email.image),
                         ),
                       ),
-                      SizedBox(width: kDefaultPadding / 2),
+                      const SizedBox(width: kDefaultPadding / 2),
                       Expanded(
                         child: Text.rich(
                           TextSpan(
@@ -60,7 +60,7 @@ class EmailCard extends StatelessWidget {
                                 style: Theme.of(context)
                                     .textTheme
                                     .bodyText2
-                                    .copyWith(
+                                    ?.copyWith(
                                       color:
                                           isActive ? Colors.white : kTextColor,
                                     ),
@@ -73,26 +73,27 @@ class EmailCard extends StatelessWidget {
                         children: [
                           Text(
                             email.time,
-                            style: Theme.of(context).textTheme.caption.copyWith(
-                                  color: isActive ? Colors.white70 : null,
-                                ),
+                            style:
+                                Theme.of(context).textTheme.caption?.copyWith(
+                                      color: isActive ? Colors.white70 : null,
+                                    ),
                           ),
-                          SizedBox(height: 5),
+                          const SizedBox(height: 5),
                           if (email.isAttachmentAvailable)
                             WebsafeSvg.asset(
-                              "assets/Icons/Paperclip.svg",
+                              "assets/icons/Paperclip.svg",
                               color: isActive ? Colors.white70 : kGrayColor,
                             )
                         ],
                       ),
                     ],
                   ),
-                  SizedBox(height: kDefaultPadding / 2),
+                  const SizedBox(height: kDefaultPadding / 2),
                   Text(
                     email.body,
                     maxLines: 2,
                     overflow: TextOverflow.ellipsis,
-                    style: Theme.of(context).textTheme.caption.copyWith(
+                    style: Theme.of(context).textTheme.caption?.copyWith(
                           height: 1.5,
                           color: isActive ? Colors.white70 : null,
                         ),
@@ -102,9 +103,9 @@ class EmailCard extends StatelessWidget {
             ).addNeumorphism(
               blurRadius: 15,
               borderRadius: 15,
-              offset: Offset(5, 5),
+              offset: const Offset(5, 5),
               topShadowColor: Colors.white60,
-              bottomShadowColor: Color(0xFF234395).withOpacity(0.15),
+              bottomShadowColor: const Color(0xFF234395).withOpacity(0.15),
             ),
             if (!email.isChecked)
               Positioned(
@@ -113,22 +114,22 @@ class EmailCard extends StatelessWidget {
                 child: Container(
                   height: 12,
                   width: 12,
-                  decoration: BoxDecoration(
+                  decoration: const BoxDecoration(
                     shape: BoxShape.circle,
                     color: kBadgeColor,
                   ),
                 ).addNeumorphism(
                   blurRadius: 4,
                   borderRadius: 8,
-                  offset: Offset(2, 2),
+                  offset: const Offset(2, 2),
                 ),
               ),
-            if (email.tagColor != null)
+            if (email.tagColor != Colors.transparent)
               Positioned(
                 left: 8,
                 top: 0,
                 child: WebsafeSvg.asset(
-                  "assets/Icons/Markup filled.svg",
+                  "assets/icons/Markup filled.svg",
                   height: 18,
                   color: email.tagColor,
                 ),
