@@ -1,20 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:outlook/components/side_menu.dart';
-import 'package:outlook/responsive.dart';
-import 'package:outlook/screens/email/email_screen.dart';
+
+import '../../components/side_menu.dart';
+import '../../responsive.dart';
+import '../email/email_screen.dart';
 import 'components/list_of_emails.dart';
 
 class MainScreen extends StatelessWidget {
+  const MainScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     // It provide us the width and height
-    Size _size = MediaQuery.of(context).size;
+    Size size = MediaQuery.of(context).size;
     return Scaffold(
       body: Responsive(
         // Let's work on our mobile part
-        mobile: ListOfEmails(),
+        mobile: const ListOfEmails(),
         tablet: Row(
-          children: [
+          children: const [
             Expanded(
               flex: 6,
               child: ListOfEmails(),
@@ -30,16 +33,16 @@ class MainScreen extends StatelessWidget {
             // Once our width is less then 1300 then it start showing errors
             // Now there is no error if our width is less then 1340
             Expanded(
-              flex: _size.width > 1340 ? 2 : 4,
-              child: SideMenu(),
+              flex: size.width > 1340 ? 2 : 4,
+              child: const SideMenu(),
             ),
             Expanded(
-              flex: _size.width > 1340 ? 3 : 5,
-              child: ListOfEmails(),
+              flex: size.width > 1340 ? 3 : 5,
+              child: const ListOfEmails(),
             ),
             Expanded(
-              flex: _size.width > 1340 ? 8 : 10,
-              child: EmailScreen(),
+              flex: size.width > 1340 ? 8 : 10,
+              child: const EmailScreen(),
             ),
           ],
         ),
